+++ lib/QECCore/src/interface.jl
@@ -117,3 +117,44 @@ Abstract type representing algorithms for computing
 the minimum distance of quantum error correction codes.
 """
 abstract type AbstractDistanceAlg end
+
+"""
+    meta_checks_x(c::AbstractCSSCode)
+
+Returns the `X`-metacheck matrix (``\\partial_{Mx}`` boundary map in
+[chain complex](https://en.wikipedia.org/wiki/Chain_complex) notation) for a CSS code.
+
+This matrix verifies validity of `X`-syndromes (`Z`-error measurements).
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`meta_checks_z`](@ref), [`meta_checks`](@ref), [`parity_matrix_x`](@ref)
+"""
+function meta_checks_x end
+
+"""
+    meta_checks_z(c::AbstractCSSCode)
+
+Returns the `Z`-metacheck matrix (``\\partial_{Mz}`` boundary map in
+[chain complex](https://en.wikipedia.org/wiki/Chain_complex) notation) for a CSS code.
+
+This matrix verifies validity of `Z`-syndromes (`X`-error measurements).
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`meta_checks_x`](@ref), [`meta_checks`](@ref), [`parity_matrix_z`](@ref)
+"""
+function meta_checks_z end
+
+"""
+    meta_checks(c::AbstractCSSCode)
+
+Returns both `X` and `Z` metacheck matrices.
+
+Only CSS codes built using chain complexes and homology have this method.
+
+See also: [`meta_checks_x`](@ref), [`meta_checks_z`](@ref)
+"""
+function meta_checks(c::AbstractCSSCode)
+    (meta_checks_x(c), meta_checks_z(c))
+end